requirements.txt
setuptools~=65.6.3
scapy~=2.5.0
click~=8.0.3
PyYAML~=5.4.1
pycryptodome~=3.17